got a workingg solution for now, I want to see if I can make it better than it is right now as it's calling a function in a file it shouldn't really be calling
